requirements.txt
beautiful==0.0.2
beautifulsoup4==4.7.1
bs4==0.0.1
Django==2.1.4
django-cors-headers==2.4.0
django-extensions==2.1.7
django-extra-fields==0.3
django-multiselectfield==0.1.8
django-rest-framework==0.1.0
djangorestframework==3.9.0
ipython==7.6.0
ipython-genutils==0.2.0
pytz==2018.7
requests==2.21.0
rules==2.0.1
selenium==3.141.0
soap==0.0.1rc0
webdriverdownloader==1.0.0.1

